don't know if header parser is universal or specific so opting to be specific
can re-use parts of this as new cases arise & refactor as more is learned. just
focusing on the specific at hand for right now, & in a way that won't require
public API changes in the future.
