+++ lib/as2/parser.rb
@@ -1,0 +1,11 @@
+module As2
+  # namespace to hold parsers for various HTTP headers & perhaps other kinds of
+  # structured content.
+  #
+  # clients should not use this directly. use the public api exposed in the
+  # top-level `As2` module itself.
+  #
+  # @api private
+  module Parser
+  end
+end
